// lib/schemas.ts
export const GetDestinationResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  country: zod.string(),
  imageUrl: zod.string().nullish(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});

export const UpdateDestinationParams = zod.object({
  destinationId: zod.coerce.number(),
});

export const UpdateDestinationBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  country: zod.string().optional(),
  imageUrl: zod.string().optional(),
  featured: zod.boolean().optional(),
});